fix: refactor pmx loader
